requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-redis==5.4.0
djangorestframework==3.16.0
idna==3.10
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
uuid==1.30
